Add blog page, first post
This adds the scaffolding for our blog (which has the clever name of
'Blog', feel free to suggest better options), and adds our first blog
post.

The blog post is the summary of the Hub talk that I would have given
yesterday had I not been sick. Figured I might as well use that video I
made for _something_.