Posting for Dev2
This is based on the posting we used for the last Dev. Only major change
is the bit at the bottom about stuff that ASR doesn't do. Better to be
upfront about that, I think.

I've made some minor text changes at the recommendation of http://www.hemingwayapp.com

I've run this through the Gender Decoder, which looks for gender bias in job
postings. It says we're 'neutral'

http://gender-decoder.katmatfield.com/results/a04c4a20-a561-45c9-b547-2cfa88886b99