Add Dev2 posting
This is mostly unchanged from our last posting

https://github.com/umn-asr/umn-asr.github.io/pull/17

I've added one new section, `What You'll Do` which outlines what things
the person will be doing in their first week, month, quarter, year. This
idea is 'borrowed' from

https://weworkremotely.com/remote-jobs/respondent-io-people-operations

The goal of this section is two-fold. First we want to give a good
picture of what the job entails, as that's a question we've gotten
before. Second, I wanted to show the that we intend to train and grow
the person we hire.