Blog UI tweaks
While going through steps of getting the blog set up for ourselves, we
made a short list of changes that immediately caught our eye.

* Moved banner to be inside the header, and removed the bottom-margin.
* Get rid of extra "Blog" link in nav
* Update subtitle copy
* Fix links to be inside headings

For now, fixing pagination by upping the pagination limit. In the future
it would be nice to possibly hardcode the link to the blog instead of
dynamically creating it, so we don't have to worry about how many posts
there are.

Links are inline elements and should not wrap block elements, so we
fixed that.

Other small layout tweaks to fix margins.

Co-authored-by: Brittany Greaner <grea0022@umn.edu>
Co-authored-by: Tony Thomas <thoma127@umn.edu>